requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
setuptools==80.9.0
celery==5.3.4
redis==5.0.1
oss2==2.18.4